Temp2

+++ app/src/main/java/com/example/root/sens/dto/Record.java
@@ -5,19 +5,19 @@ import com.example.root.sens.dto.ActivityCategories;
 import io.realm.RealmObject;
 
 public class Record extends RealmObject {
-    private int value;
+    private float value;
     private String type;
     public Record(){}
-    public Record(int value, String type) {
+    public Record(float value, String type) {
         this.value = value;
         this.type = type;
     }
 
-    public int getValue() {
+    public float getValue() {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(float value) {
         this.value = value;
     }
